src/data_analysis/requirements-backend.txt
nltk
Flask==2.1.2
Jinja2==3.1.2
Werkzeug==2.1.1
psycopg2-binary
matplotlib==3.4.2
